REFACTOR: move is_htmx_request to dependencies
