+++ src/air/__init__.py
@@ -7,7 +7,8 @@ from .applications import Air as Air
 from .background import BackgroundTasks as BackgroundTasks
 from .forms import AirField as AirField, AirForm as AirForm
 from .middleware import SessionMiddleware as SessionMiddleware
-from .requests import is_htmx_request as is_htmx_request, Request as Request
+from .requests import Request as Request
+from .dependencies import is_htmx_request as is_htmx_request
 from .responses import (
     AirResponse as AirResponse,
     SSEResponse as SSEResponse,
